+++ include/dag/layer/conv.h
@@ -13,8 +13,8 @@ namespace dl {
 namespace nn {
 class Conv2D : public LayerBase {
   public:
-    explicit Conv2D(int64_t in_channels, int64_t out_channels, int64_t kernel_size,
-                    int64_t stride = 1, int64_t padding = 0, bool bias = true,
+    explicit Conv2D(int in_channels, int out_channels, int kernel_size,
+                    int stride = 1, int padding = 0, bool bias = true,
                     const char *padding_mode = "zeros")
         : mInChannels(in_channels)
         , mOutChannels(out_channels)
@@ -29,7 +29,7 @@ class Conv2D : public LayerBase {
             CreateNode({mOutChannels, mInChannels, mKernelSize, mKernelSize}, true);
         Random(mWeight->tensor()->data(), mWeight->tensor()->size());
 
-        LOG_INFO("\n\tconv args:\n\t\tin_channels:%lu\n\t\tout_channels:%lu\n\t\tkernel_size:%lu\n\t\tstride:%lu\n\t\tpadding:%lu\n\t\tpaddingmode:%s\n\t\tbias:%s", mInChannels, mOutChannels, mKernelSize, mStride, mPadding, mPaddingMode.c_str(), bias ? "true": "false");
+        LOG_INFO("\n\tconv args:\n\t\tin_channels:%d\n\t\tout_channels:%d\n\t\tkernel_size:%d\n\t\tstride:%d\n\t\tpadding:%d\n\t\tpaddingmode:%s\n\t\tbias:%s", mInChannels, mOutChannels, mKernelSize, mStride, mPadding, mPaddingMode.c_str(), bias ? "true": "false");
         if (mBiasFlag) {
             mBias = CreateNode({mOutChannels}, true);
             Random(mBias->tensor()->data(), mBias->tensor()->size());
@@ -50,11 +50,11 @@ class Conv2D : public LayerBase {
     }
 
   private:
-    int64_t      mInChannels;
-    int64_t      mOutChannels;
-    int64_t      mKernelSize;
-    int64_t      mStride;
-    int64_t      mPadding;
+    int      mInChannels;
+    int      mOutChannels;
+    int      mKernelSize;
+    int      mStride;
+    int      mPadding;
     bool        mBiasFlag;
     std::string mPaddingMode;
     DataNode *  mWeight;
