+++ include/dag/operator/crossentropy.h
@@ -14,7 +14,7 @@
 namespace dl {
 class CrossEntropyImpl : public OperatorNodeBase {
   public:
-    CrossEntropyImpl(int64_t uid)
+    CrossEntropyImpl(int uid)
         : OperatorNodeBase(uid) {}
     ~CrossEntropyImpl() = default;
 
@@ -24,11 +24,11 @@ class CrossEntropyImpl : public OperatorNodeBase {
         const Tensor *logits = inps[0]; // (n, c)
         const Tensor *labels = inps[1]; // (n)
         LOG_DEBUG("label shape:%s", FormatShape(labels->shape()).c_str());
-        int64_t num_classes = logits->shape()[1];
-        for (int64_t n = 0; n < logits->shape()[0]; ++n) {
-            int64_t idx = static_cast<int64_t>(labels->data()[n]);
+        int num_classes = logits->shape()[1];
+        for (int n = 0; n < logits->shape()[0]; ++n) {
+            int idx = static_cast<int>(labels->data()[n]);
             if(idx >= num_classes){
-                LOG_ERROR("idx %lu out of range %lu", idx, num_classes);
+                LOG_ERROR("idx %d out of range %d", idx, num_classes);
             }
             out->data()[n] =
                 logits->data()[n * num_classes + idx];
@@ -38,9 +38,9 @@ class CrossEntropyImpl : public OperatorNodeBase {
 
     void backward(const Tensor *diff, std::vector<Tensor *> &grads) override {
         Tensor *grad_logits = grads[0];
-        int64_t num_classes = grad_logits->shape()[1];
-        for(int64_t n = 0; n < diff->shape()[0]; ++n){
-            grad_logits->data()[n * num_classes + static_cast<int64_t>(mLabelIdx.data()[n])] = diff->data()[n];
+        int num_classes = grad_logits->shape()[1];
+        for(int n = 0; n < diff->shape()[0]; ++n){
+            grad_logits->data()[n * num_classes + static_cast<int>(mLabelIdx.data()[n])] = diff->data()[n];
         }
     }
     Shape inferenceShape(const std::vector<const Tensor *> &inps) override {
